repo-step-3/pyproject.toml
[project]
name = "litellm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "nest-asyncio>=1.6.0",
    "python-dotenv>=1.1.0",
]

[dependency-groups]
dev = [
    
]

[tool.uv.sources]
litellm = { workspace = true }
